Add test case for the problem with concurrent eviction
