+++ tortoise/rerun_test.go
@@ -7,7 +7,9 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/spacemeshos/go-spacemesh/common/types"
+	"github.com/spacemeshos/go-spacemesh/log/logtest"
 	"github.com/spacemeshos/go-spacemesh/mesh"
+	"github.com/spacemeshos/go-spacemesh/tortoise/sim"
 )
 
 func TestRerunAndRevert(t *testing.T) {
@@ -70,4 +72,69 @@ func TestRerunAndRevert(t *testing.T) {
 	isValid, err = mdb.ContextualValidity(block1ID)
 	r.NoError(err)
 	r.False(isValid)
+}
+
+func TestRerunMissingLayer(t *testing.T) {
+	ctx := context.Background()
+	s := sim.New(sim.WithLayerSize(defaultTestLayerSize))
+	s.Setup()
+
+	cfg := defaultConfig(t, s.State.MeshDB, s.State.AtxDB)
+	tortoise := NewVerifyingTortoise(ctx, cfg)
+	layers := make([]types.LayerID, 20)
+	for i := range layers {
+		layers[i] = s.Next()
+	}
+	missing := layers[10]
+	for _, lid := range layers {
+		if lid == missing {
+			continue
+		}
+		tortoise.HandleIncomingLayer(context.TODO(), lid)
+	}
+	// TODO(dshulyak) implement public interface to enforce rerun
+	require.NoError(t, tortoise.rerun(context.TODO()))
+	oldpbase, newpbase, reverted := tortoise.HandleIncomingLayer(context.TODO(), s.Next())
+	require.True(t, reverted)
+	require.Equal(t, missing.Sub(2), oldpbase)
+	require.Equal(t, layers[len(layers)-1], newpbase)
+	for _, lid := range layers {
+		bids, err := s.State.MeshDB.LayerBlockIds(lid)
+		require.NoError(t, err)
+		for _, bid := range bids {
+			validity, err := s.State.MeshDB.ContextualValidity(bid)
+			require.NoError(t, err)
+			require.True(t, validity)
+		}
+	}
+}
+
+func TestRerunEvictConcurrent(t *testing.T) {
+	ctx := context.Background()
+	s := sim.New(sim.WithLayerSize(defaultTestLayerSize))
+	s.Setup()
+
+	cfg := defaultConfig(t, s.State.MeshDB, s.State.AtxDB)
+	tortoise := NewVerifyingTortoise(ctx, cfg)
+
+	for i := 0; i < int(cfg.WindowSize); i++ {
+		tortoise.HandleIncomingLayer(ctx, s.Next())
+	}
+	require.NoError(t, tortoise.Persist(ctx))
+	require.NoError(t, tortoise.rerun(ctx))
+	for i := 0; i < int(cfg.WindowSize); i++ {
+		// have to use private trtl to simulate concurrent layers without replacing tortoise instance
+		tortoise.trtl.HandleIncomingLayer(ctx, s.Next())
+	}
+	require.NoError(t, tortoise.Persist(ctx))
+	last := s.Next()
+	tortoise.HandleIncomingLayer(ctx, last)
+	require.NoError(t, tortoise.Persist(ctx))
+
+	st := state{log: logtest.New(t), db: cfg.Database}
+	require.NoError(t, st.Recover())
+	require.Equal(t, last.Sub(1), st.Verified)
+	for lid := range st.BlockOpinionsByLayer {
+		require.True(t, lid.After(st.LastEvicted))
+	}
 }
